Rename "multi_source_registry.rs"
